requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.12.0
python-dotenv==1.0.1
newspaper3k==0.2.8
lxml==5.1.0
python-dateutil==2.8.2